requirements.txt
aiogram==3.3.0
asyncpg==0.29.0
python-dotenv==1.0.0
psycopg2-binary==2.9.7
schedule==1.2.0
aiofiles==23.2.1
aiohttp==3.9.1